+++ cmd/leases.go
@@ -12,7 +12,7 @@ var loginOpenBrowser bool
 var loginPrintCreds bool
 var loginProfile string
 
-var principleID string
+var principalID string
 var budgetAmount float64
 var budgetCurrency string
 var email []string
@@ -30,23 +30,23 @@ func init() {
 	leasesListCmd.Flags().Int64VarP(&pagLimit, "limit", "l", defaultPagLiimt, "Max number of leases to return at once. Will include url to next page if there is one.")
 	leasesListCmd.Flags().StringVarP(&nextAcctID, "next-account-id", "", "", "Account ID with which to begin the scan operation. This is used to traverse through paginated results.")
 	leasesListCmd.Flags().StringVarP(&nextPrincipalID, "next-principal-id", "", "", "Principal ID with which to begin the scan operation. This is used to traverse through paginated results.")
-	leasesListCmd.Flags().StringVarP(&principleID, "principle-id", "p", "", "Principle ID of a user")
+	leasesListCmd.Flags().StringVarP(&principalID, "principal-id", "p", "", "Principle ID of a user")
 	leasesListCmd.Flags().StringVarP(&leaseStatus, "status", "s", "", "Lease status")
 	leasesCmd.AddCommand(leasesListCmd)
 
-	leasesCreateCmd.Flags().StringVarP(&principleID, "principle-id", "p", "", "Principle ID for the user of the leased account")
+	leasesCreateCmd.Flags().StringVarP(&principalID, "principal-id", "p", "", "Principle ID for the user of the leased account")
 	leasesCreateCmd.Flags().Float64VarP(&budgetAmount, "budget-amount", "b", 0, "The leased accounts budget amount")
 	leasesCreateCmd.Flags().StringVarP(&budgetCurrency, "budget-currency", "c", "USD", "The leased accounts budget currency")
 	leasesCreateCmd.Flags().StringArrayVarP(&email, "email", "e", nil, "The email address that budget notifications will be sent to")
-	leasesCreateCmd.MarkFlagRequired("principle-id")
+	leasesCreateCmd.MarkFlagRequired("principal-id")
 	leasesCreateCmd.MarkFlagRequired("budget-amount")
 	leasesCreateCmd.MarkFlagRequired("budget-currency")
 	leasesCreateCmd.MarkFlagRequired("email")
 	leasesCmd.AddCommand(leasesCreateCmd)
 
-	leasesEndCmd.Flags().StringVarP(&principleID, "principle-id", "p", "", "Principle ID for the user of the leased account")
+	leasesEndCmd.Flags().StringVarP(&principalID, "principal-id", "p", "", "Principle ID for the user of the leased account")
 	leasesEndCmd.Flags().StringVarP(&accountID, "account-id", "a", "", "Account ID associated with the lease you wish to end")
-	leasesEndCmd.MarkFlagRequired("principle-id")
+	leasesEndCmd.MarkFlagRequired("principal-id")
 	leasesEndCmd.MarkFlagRequired("account-id")
 	leasesCmd.AddCommand(leasesEndCmd)
 
@@ -77,7 +77,7 @@ var leasesListCmd = &cobra.Command{
 	Short: "List leases using various query filters.",
 	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
-		service.ListLeases(acctID, principleID, nextAcctID, nextPrincipalID, leaseStatus, pagLimit)
+		service.ListLeases(acctID, principalID, nextAcctID, nextPrincipalID, leaseStatus, pagLimit)
 	},
 }
 
@@ -86,7 +86,7 @@ var leasesCreateCmd = &cobra.Command{
 	Short: "Create a lease.",
 	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
-		service.CreateLease(principleID, budgetAmount, budgetCurrency, email)
+		service.CreateLease(principalID, budgetAmount, budgetCurrency, email)
 	},
 }
 
@@ -95,7 +95,7 @@ var leasesEndCmd = &cobra.Command{
 	Short: "Cause a lease to immediately expire",
 	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
-		service.EndLease(accountID, principleID)
+		service.EndLease(accountID, principalID)
 	},
 }