+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## vNext
+
+**BREAKING CHANGES**
+- Fix `--principle-id` flag to be `principal-id`
+
 ## v0.1.0
 
 Initial release